requirements.txt
fusepy
libtorrent==2.0.9
torrentstream